+++ src/main/java/pl/pawz/zelbet/ULS/ExtensionAsymmetricReinforcement.java
@@ -5,15 +5,15 @@ import pl.pawz.zelbet.BasicValuesPillars;
 import pl.pawz.zelbet.PolynomialSolver;
 
 public class ExtensionAsymmetricReinforcement {
-    private float nEd;
+    private double nEd;
     private double epsilonCu3;
     private double fCd;
     private double fYd;
     private double etaConcrete;
     private double lambdaConcrete;
     private double dDimension;
-    private float bDimension;
-    private float a2;
+    private double bDimension;
+    private double a2;
     private int E_S;
     private double xMinusMinYd;
     private double xMinYd;
@@ -26,8 +26,8 @@ public class ExtensionAsymmetricReinforcement {
     private double aS2;
 
 
-    public ExtensionAsymmetricReinforcement(float nEd, float mEd, double fCk, double fYk, float bDimension,
-                                            float hDimension, float a1, float a2) {
+    public ExtensionAsymmetricReinforcement(double nEd, double mEd, double fCk, double fYk, double bDimension,
+                                            double hDimension, double a1, double a2) {
         this.nEd = nEd;
         this.epsilonCu3 = BasicValues.epsilonCu3Value(fCk);
         this.fCd = BasicValues.fCdValue(fCk);
